Refactored Repository Class into DroneRepository and CameraRepository classes
